requirements.txt
django==3.0.6
plaid-python==4.0.0
